Cautiously expose new user facing API through experimental file
