+++ db/manifest_ops.cc
@@ -6,11 +6,7 @@
 
 #include "db/manifest_ops.h"
 
-#include "db/version_edit_handler.h"
 #include "file/filename.h"
-#include "file/sequence_file_reader.h"
-#include "rocksdb/env.h"
-#include "rocksdb/file_checksum.h"
 
 namespace ROCKSDB_NAMESPACE {
 
@@ -46,63 +42,6 @@ Status GetCurrentManifestPath(const std::string& dbname, FileSystem* fs,
   }
   manifest_path->append(fname);
   return Status::OK();
-}
-
-Status GetFileChecksumsFromCurrentManifest(FileSystem* fs,
-                                           const std::string& dbname,
-                                           FileChecksumList* checksum_list) {
-  std::string manifest_path;
-  uint64_t manifest_file_number;
-  Status s = GetCurrentManifestPath(dbname, fs, true /* is_retry */,
-                                    &manifest_path, &manifest_file_number);
-  if (!s.ok()) {
-    return s;
-  }
-
-  if (checksum_list == nullptr) {
-    return Status::InvalidArgument("checksum_list is nullptr");
-  }
-  assert(checksum_list);
-
-  const ReadOptions read_options(Env::IOActivity::kReadManifest);
-  checksum_list->reset();
-
-  std::unique_ptr<SequentialFileReader> file_reader;
-  {
-    std::unique_ptr<FSSequentialFile> file;
-    s = fs->NewSequentialFile(manifest_path,
-                              fs->OptimizeForManifestRead(FileOptions()), &file,
-                              nullptr /* dbg */);
-    if (!s.ok()) {
-      return s;
-    }
-    file_reader.reset(new SequentialFileReader(std::move(file), manifest_path));
-  }
-
-  struct LogReporter : public log::Reader::Reporter {
-    Status* status_ptr;
-    void Corruption(size_t /*bytes*/, const Status& st) override {
-      if (status_ptr->ok()) {
-        *status_ptr = st;
-      }
-    }
-  } reporter;
-  reporter.status_ptr = &s;
-  log::Reader reader(nullptr, std::move(file_reader), &reporter,
-                     true /* checksum */, 0 /* log_number */);
-
-  uint64_t manifest_file_size;
-  IOOptions opts;
-  fs->GetFileSize(manifest_path, opts, &manifest_file_size, nullptr);
-
-  FileChecksumRetriever retriever(read_options, manifest_file_size,
-                                  *checksum_list);
-  retriever.Iterate(reader, &s);
-  assert(!retriever.status().ok() ||
-         manifest_file_size == std::numeric_limits<uint64_t>::max() ||
-         reader.LastRecordEnd() == manifest_file_size);
-
-  return retriever.status();
 }
 
 }  // namespace ROCKSDB_NAMESPACE
